Allow matcher to accept results object
