+++ src/index.ts
@@ -1,2 +1,3 @@
 import matchers from './matchers'
+export { runAxe } from './axe'
 export default matchers
