+++ src/types/index.ts
@@ -1,0 +1,6 @@
+import type { RunOptions } from 'axe-core'
+
+export interface MatcherOptions extends RunOptions {
+  timeout?: number
+  filename?: string
+}
